first cut separation
