+++ src/index.d.ts
@@ -1,0 +1,11 @@
+type CodeSnippetParams = { [k: string]: string };
+type CodeSnippetParamsMap = {
+  raw: CodeSnippetParams;
+  params: CodeSnippetParams;
+};
+
+interface PluginDataAndParams {
+  codeArray: string[];
+  pluginDataArray: CodegenResult[];
+  nodeType: string;
+}
