+++ src/utils.ts
@@ -1,0 +1,34 @@
+export type StringFilter =
+  | "hyphen"
+  | "camel"
+  | "constant"
+  | "pascal"
+  | "raw"
+  | "snake";
+
+export function formatString(
+  string: string,
+  rawString: string,
+  filter: StringFilter
+) {
+  if (!filter) filter = "hyphen";
+  const splitString = string.split("-");
+  const capitalize = (s: string) => s.charAt(0).toUpperCase() + s.substring(1);
+  switch (filter) {
+    case "camel":
+      return splitString
+        .map((word, i) => (i === 0 ? word : capitalize(word)))
+        .join("");
+    case "constant":
+      return splitString.join("_").toUpperCase();
+    case "hyphen":
+      return splitString.join("-").toLowerCase();
+    case "pascal":
+      return splitString.map(capitalize).join("");
+    case "raw":
+      return rawString;
+    case "snake":
+      return splitString.join("_").toLowerCase();
+  }
+  return splitString.join(" ");
+}
